Refactor game loop into modular update and render
