+++ src/main.js
@@ -1,3 +1,20 @@
+import { startGameLoop } from "./game/loop.js";
+import {
+  createUpdate,
+  wrapOffset,
+  getWorldOffsetsAround as computeWorldOffsetsAround,
+  getWorldWrapOffsetsForView as computeWorldWrapOffsetsForView,
+  isNear
+} from "./game/update.js";
+import {
+  createRender,
+  queueEffect,
+  createJumpDustEffect,
+  createCrystalBurstEffect,
+  createPortalActivationEffect,
+  createScreenShakeEffect
+} from "./game/render.js";
+
 // Provide minimal runtime shims before the rest of the lobby bootstraps. These
 // ensure the game can execute on browsers that predate modern globals but still
 // support ES modules.
@@ -5040,6 +5057,16 @@ if (activeAccountLobbyLayout) {
 let worldWrapWidth = computeWorldWrapWidth();
 const worldWrapOffsetsScratch = [0, 0, 0];
 const worldWrapRenderOffsetsScratch = [];
+const getWorldOffsetsAround = (value) =>
+  computeWorldOffsetsAround(value, worldWrapWidth, worldWrapOffsetsScratch);
+const getWorldWrapOffsetsForView = (cameraOffset, viewportWidth = viewport.width, margin = 0) =>
+  computeWorldWrapOffsetsForView(
+    cameraOffset,
+    viewportWidth,
+    worldWrapWidth,
+    worldWrapRenderOffsetsScratch,
+    margin
+  );
 
 function computeWorldWrapWidth() {
   const extents = [];
@@ -5793,692 +5820,151 @@ ui.updateCrystals(0, crystals.length);
 ui.refresh(playerStats);
 refreshMissionDisplay();
 
-const FIXED_TIMESTEP = 16.666;
-const MAX_FRAME_TIME = 100;
-const MAX_ACCUMULATED_TIME = FIXED_TIMESTEP * 5;
+const update = createUpdate({
+  state: {
+    player,
+    playerStats,
+    portal,
+    crystals,
+    interactables,
+    platforms,
+    groundY,
+    portalRequiredLevel,
+    getPortalCharge: () => portalCharge,
+    setPortalCharge: (value) => {
+      portalCharge = value;
+    },
+    isPortalCharged: () => portalCharged,
+    setPortalCharged: (value) => {
+      portalCharged = value;
+    },
+    isPortalCompleted: () => portalCompleted,
+    setPortalCompleted: (value) => {
+      portalCompleted = value;
+    },
+    getCrystalRunStartTime: () => crystalRunStartTime,
+    setCrystalRunStartTime: (value) => {
+      crystalRunStartTime = value;
+    },
+    getParallaxScroll: () => parallaxScroll,
+    setParallaxScroll: (value) => {
+      parallaxScroll = value;
+    },
+    getLayoutEditor: () => layoutEditor,
+    isMiniGameActive: () => miniGameActive,
+    getCameraScrollX: () => cameraScrollX,
+    getRenderScale: () => renderScale,
+    viewport
+  },
+  world: {
+    getWorldOffsetsAround,
+    createWorldInstance,
+    isNear,
+    getInteractableDisplayName
+  },
+  movementHint: {
+    isVisible: () => movementHintVisible,
+    setVisible: (value) => {
+      movementHintVisible = value;
+    },
+    isAcknowledged: () => movementHintAcknowledged,
+    wasShownThisSession: () => movementHintShownThisSession,
+    setShownThisSession: (value) => {
+      movementHintShownThisSession = value;
+    },
+    getIdleTime: () => movementHintIdleTime,
+    setIdleTime: (value) => {
+      movementHintIdleTime = value;
+    },
+    dismiss: dismissMovementHint
+  },
+  input: {
+    keys,
+    wasKeyJustPressed
+  },
+  ui,
+  audio,
+  notifyFirstSessionChecklist,
+  showMessage,
+  gainExperience,
+  completeMission,
+  openMiniGame,
+  effects: {
+    queueEffect,
+    createJumpDustEffect,
+    createCrystalBurstEffect,
+    createPortalActivationEffect,
+    createScreenShakeEffect
+  },
+  updateCameraScrollPosition,
+  constants: {
+    MOVEMENT_HINT_IDLE_THRESHOLD,
+    PARALLAX_IDLE_SCROLL_SPEED,
+    CHECKLIST_EVENTS,
+    CRYSTAL_SPRINT_TARGET_MS
+  },
+  utils: { clamp }
+});
 
-let lastTimestamp = performance.now();
-let accumulatedFrameTime = 0;
-requestAnimationFrame(loop);
-
-function loop(timestamp) {
-  let frameTime = timestamp - lastTimestamp;
-  if (!Number.isFinite(frameTime) || frameTime < 0) {
-    frameTime = FIXED_TIMESTEP;
+const renderState = {
+  getRenderScale: () => renderScale,
+  getDevicePixelScale: () => devicePixelScale,
+  getCameraScrollX: () => cameraScrollX,
+  viewport,
+  getWorldWrapOffsetsForView: (cameraOffset, viewportWidth = viewport.width, margin = 0) =>
+    computeWorldWrapOffsetsForView(
+      cameraOffset,
+      viewportWidth,
+      worldWrapWidth,
+      worldWrapRenderOffsetsScratch,
+      margin
+    ),
+  wrapOffset,
+  parallaxLayers,
+  getParallaxScroll: () => parallaxScroll,
+  groundY,
+  getFallbackBackgroundGradient,
+  platformSprite,
+  platforms,
+  isWorldInstanceVisible,
+  createWorldInstance,
+  portal,
+  isPortalCharged: () => portalCharged,
+  interactables,
+  crystals,
+  player,
+  getLayoutEditor: () => layoutEditor,
+  ui,
+  playerSpriteState,
+  playerAppearance,
+  promptFont,
+  getPromptMetrics,
+  sprites: {
+    chestSprite,
+    arcadeSprite,
+    fountainSprite,
+    guideSprite,
+    crystalSprite
   }
-  frameTime = Math.min(frameTime, MAX_FRAME_TIME);
-  lastTimestamp = timestamp;
+};
 
-  accumulatedFrameTime = Math.min(accumulatedFrameTime + frameTime, MAX_ACCUMULATED_TIME);
+const renderFrame = createRender(renderState);
 
-  frameJustPressed.clear();
-  for (const code of justPressed) {
-    frameJustPressed.add(code);
-  }
-  let processedInput = false;
-
-  while (accumulatedFrameTime >= FIXED_TIMESTEP) {
-    currentJustPressed = processedInput ? emptyJustPressedSet : frameJustPressed;
-    update(FIXED_TIMESTEP);
-    accumulatedFrameTime -= FIXED_TIMESTEP;
-    processedInput = true;
-  }
-
-  currentJustPressed = frameJustPressed;
-  render(timestamp);
-
-  if (processedInput) {
-    justPressed.clear();
-  }
-
-  requestAnimationFrame(loop);
-}
-
-function update(delta) {
-  const portalWasCharged = portalCharged;
-  const previousY = player.y;
-
-  if (miniGameActive) {
-    player.vx = 0;
-    player.vy = 0;
-    player.onGround = true;
-    ui.setPrompt("");
-    dismissMovementHint();
-    return;
-  }
-
-  if (layoutEditor && typeof layoutEditor.isActive === "function" && layoutEditor.isActive()) {
-    player.vx = 0;
-    player.vy = 0;
-    player.onGround = true;
-    ui.setPrompt("");
-    dismissMovementHint();
-    return;
-  }
-
-  const moveLeft = keys.has("ArrowLeft") || keys.has("KeyA");
-  const moveRight = keys.has("ArrowRight") || keys.has("KeyD");
-  const jumpPressed =
-    wasKeyJustPressed("Space") ||
-    wasKeyJustPressed("ArrowUp") ||
-    wasKeyJustPressed("KeyW");
-
-  const jumpHeld = keys.has("Space") || keys.has("ArrowUp") || keys.has("KeyW");
-  const horizontalHeld =
-    moveLeft ||
-    moveRight ||
-    keys.has("ArrowLeft") ||
-    keys.has("ArrowRight") ||
-    keys.has("KeyA") ||
-    keys.has("KeyD");
-  const hasMovementOrJumpInput = horizontalHeld || jumpHeld || jumpPressed;
-
-  if (hasMovementOrJumpInput) {
-    if (movementHintVisible) {
-      dismissMovementHint();
-    } else {
-      movementHintIdleTime = 0;
-    }
-  } else if (!movementHintAcknowledged && !movementHintShownThisSession) {
-    movementHintIdleTime = Math.min(
-      movementHintIdleTime + delta,
-      MOVEMENT_HINT_IDLE_THRESHOLD * 2
-    );
-  }
-
-  const acceleration = 0.35 * (delta / 16.666);
-  const maxSpeed = 4.2;
-  const friction = 0.82;
-  const gravity = 0.52 * (delta / 16.666);
-
-  if (moveLeft && !moveRight) {
-    player.vx = Math.max(player.vx - acceleration, -maxSpeed);
-    player.direction = -1;
-  } else if (moveRight && !moveLeft) {
-    player.vx = Math.min(player.vx + acceleration, maxSpeed);
-    player.direction = 1;
-  } else {
-    player.vx *= friction;
-    if (Math.abs(player.vx) < 0.01) {
-      player.vx = 0;
-    }
-  }
-
-  if (jumpPressed && player.onGround) {
-    player.vy = -10.8;
-    player.onGround = false;
-    audio.playEffect("jump");
-    queueEffect(
-      createJumpDustEffect(player.x + player.width / 2, player.y + player.height)
-    );
-  }
-
-  player.vy += gravity;
-  player.x += player.vx * (delta / 16.666);
-  player.y += player.vy * (delta / 16.666);
-  player.onGround = false;
-
-  if (player.y + player.height >= groundY) {
-    player.y = groundY - player.height;
-    player.vy = 0;
-    player.onGround = true;
-  }
-
-  if (PARALLAX_IDLE_SCROLL_SPEED === 0) {
-    parallaxScroll = 0;
-  } else {
-    parallaxScroll += PARALLAX_IDLE_SCROLL_SPEED * (delta / 16.666);
-    if (!Number.isFinite(parallaxScroll)) {
-      parallaxScroll = 0;
-    } else if (parallaxScroll > 10000 || parallaxScroll < -10000) {
-      parallaxScroll = wrapOffset(parallaxScroll, 10000);
-      if (parallaxScroll > 5000) {
-        parallaxScroll -= 10000;
-      }
-    }
-  }
-
-  updateCameraScrollPosition();
-
-  if (
-    !movementHintAcknowledged &&
-    !movementHintVisible &&
-    !movementHintShownThisSession &&
-    movementHintIdleTime >= MOVEMENT_HINT_IDLE_THRESHOLD
-  ) {
-    movementHintShownThisSession = true;
-    movementHintVisible = true;
-  }
-
-  const platformOffsets = getWorldOffsetsAround(player.x);
-  for (const offset of platformOffsets) {
-    for (const platform of platforms) {
-      const platformX = platform.x + offset;
-      const isAbovePlatform = previousY + player.height <= platform.y;
-      const isWithinX =
-        player.x + player.width > platformX &&
-        player.x < platformX + platform.width;
-
-      if (player.vy >= 0 && isAbovePlatform && isWithinX) {
-        const bottom = player.y + player.height;
-        if (bottom >= platform.y && bottom <= platform.y + platform.height + 4) {
-          player.y = platform.y - player.height;
-          player.vy = 0;
-          player.onGround = true;
-        }
-      }
+startGameLoop(
+  (delta) => update(delta),
+  (timestamp) => renderFrame(ctx, timestamp),
+  {
+    justPressed,
+    frameJustPressed,
+    emptyJustPressedSet,
+    setCurrentJustPressed: (set) => {
+      currentJustPressed = set;
+    },
+    clearJustPressed: () => {
+      justPressed.clear();
     }
   }
-
-  let promptText = "";
-  let promptTarget = null;
-
-  const crystalOffsets = getWorldOffsetsAround(player.x);
-  for (const offset of crystalOffsets) {
-    for (const crystal of crystals) {
-      if (crystal.collected) continue;
-
-      const crystalX = crystal.x + offset;
-      const overlapX =
-        player.x + player.width > crystalX - crystal.radius &&
-        player.x < crystalX + crystal.radius;
-      const overlapY =
-        player.y + player.height > crystal.y - crystal.radius &&
-        player.y < crystal.y + crystal.radius;
-
-      if (overlapX && overlapY) {
-        crystal.collected = true;
-        audio.playEffect("crystal");
-        queueEffect(createCrystalBurstEffect(crystalX, crystal.y));
-        const hadNoCharge = portalCharge === 0;
-        notifyFirstSessionChecklist(CHECKLIST_EVENTS.CRYSTAL_COLLECTED, {
-          first: hadNoCharge,
-          crystal
-        });
-        const now =
-          typeof performance !== "undefined" && typeof performance.now === "function"
-            ? performance.now()
-            : Date.now();
-        if (hadNoCharge) {
-          crystalRunStartTime = now;
-        }
-        portalCharge = Math.min(portalCharge + 1, crystals.length);
-        ui.updateCrystals(portalCharge, crystals.length);
-        const fullyCharged = portalCharge === crystals.length;
-        if (fullyCharged) {
-          if (!portalWasCharged) {
-            audio.playEffect("portalCharge");
-          }
-          portalCharged = true;
-        }
-        const leveledUpFromCrystal = gainExperience(60);
-        const levelUpNotices = new Set();
-        if (!fullyCharged && leveledUpFromCrystal) {
-          levelUpNotices.add(`Level up! You reached level ${playerStats.level}.`);
-        }
-        const messageParts = [];
-        if (fullyCharged) {
-          const portalReady = playerStats.level >= portalRequiredLevel;
-          messageParts.push(
-            portalReady
-              ? "The final crystal ignites the portal! +60 EXP. Return and press E to travel onward."
-              : `The final crystal ignites the portal! +60 EXP. Reach Level ${portalRequiredLevel} before entering.`
-          );
-        } else {
-          messageParts.push("Crystal energy surges through you! +60 EXP.");
-        }
-
-        const highlightParts = [];
-        if (fullyCharged) {
-          const chargeMission = completeMission("mission-crystal-charge");
-          if (chargeMission.completed) {
-            const xpAward = chargeMission.mission?.xp ?? 0;
-            const title = chargeMission.mission?.title ?? "Portal Mission";
-            highlightParts.push(`Mission Complete: ${title}! +${xpAward} EXP.`);
-            if (chargeMission.leveledUp) {
-              levelUpNotices.add(`Level up! You reached level ${playerStats.level}.`);
-            }
-          } else if (chargeMission.locked) {
-            const required = Math.max(1, chargeMission.mission?.requiredLevel ?? 1);
-            highlightParts.push(`Train to Level ${required} to log the portal charge mission.`);
-          }
-
-          const runDurationMs = crystalRunStartTime > 0 ? now - crystalRunStartTime : 0;
-          if (runDurationMs > 0) {
-            const runSeconds = Math.max(0, Math.round(runDurationMs / 1000));
-            highlightParts.push(`Crystal run time: ${runSeconds}s.`);
-            if (runDurationMs <= CRYSTAL_SPRINT_TARGET_MS) {
-              const sprintMission = completeMission("mission-crystal-sprint");
-              if (sprintMission.completed) {
-                const xpAward = sprintMission.mission?.xp ?? 0;
-                const title = sprintMission.mission?.title ?? "Crystal Sprint";
-                highlightParts.push(`Mission Complete: ${title}! +${xpAward} EXP.`);
-                if (sprintMission.leveledUp) {
-                  levelUpNotices.add(`Level up! You reached level ${playerStats.level}.`);
-                }
-              } else if (sprintMission.locked) {
-                const required = Math.max(1, sprintMission.mission?.requiredLevel ?? 1);
-                highlightParts.push(
-                  `Reach Level ${required} to record sprint times on the mission board.`
-                );
-              }
-            }
-          }
-          crystalRunStartTime = 0;
-        }
-
-        for (const notice of levelUpNotices) {
-          highlightParts.push(notice);
-        }
-
-        const message = [...messageParts, ...highlightParts].join(" ");
-        showMessage(
-          {
-            text: message,
-            author: "Mission Log",
-            channel: "mission"
-          },
-          fullyCharged ? 5800 : 4200
-        );
-        break;
-      }
-    }
-  }
-
-  const interactableOffsets = getWorldOffsetsAround(player.x);
-  for (const offset of interactableOffsets) {
-    for (const interactable of interactables) {
-      const instance = createWorldInstance(interactable, offset);
-      const near = isNear(player, instance, 24);
-      if (!near) {
-        continue;
-      }
-
-      if (interactable.type === "bulletin") {
-        promptText = "Press E to review the bulletin board";
-        promptTarget = instance;
-        if (wasKeyJustPressed("KeyE")) {
-          const result = completeMission(interactable.missionId);
-          audio.playEffect("dialogue");
-          if (result.completed) {
-            const xpAward = result.mission?.xp ?? 0;
-            let message =
-              `You craft a triumphant enlistment post for the cosmos. +${xpAward} EXP.`;
-            if (result.leveledUp) {
-              message += ` Level up! You reached level ${playerStats.level}.`;
-            }
-            showMessage(
-              { text: message, author: "Mission Log", channel: "mission" },
-              5600
-            );
-          } else if (result.locked) {
-            const required = Math.max(1, result.mission?.requiredLevel ?? 1);
-            showMessage(
-              {
-                text: `Level ${required} required before the bulletin board can log your official broadcast.`,
-                author: "Mission Log",
-                channel: "mission"
-              },
-              4200
-            );
-          } else if (result.alreadyComplete) {
-            showMessage(
-              {
-                text: "Your arrival announcement already glows across the bulletin board.",
-                author: "Mission Log",
-                channel: "mission"
-              },
-              4200
-            );
-          } else {
-            showMessage(
-              {
-                text: "Mission updates shimmer across the bulletin board display.",
-                author: "Mission Log",
-                channel: "mission"
-              },
-              3800
-            );
-          }
-        }
-      } else if (interactable.type === "chest") {
-        promptText = "Press E to open the chest";
-        promptTarget = instance;
-        if (wasKeyJustPressed("KeyE")) {
-          if (!interactable.opened) {
-            interactable.opened = true;
-            playerStats.hp = clamp(playerStats.hp + 12, 0, playerStats.maxHp);
-            ui.refresh(playerStats);
-            audio.playEffect("chestOpen");
-            showMessage(
-              {
-                text: "You found herbal tonics! HP restored.",
-                author: "Mission Log",
-                channel: "mission"
-              },
-              3600
-            );
-          } else {
-            audio.playEffect("dialogue");
-            showMessage(
-              {
-                text: "The chest is empty now, but still shiny.",
-                author: "Mission Log",
-                channel: "mission"
-              },
-              2800
-            );
-          }
-        }
-      } else if (interactable.type === "arcade") {
-        promptText = "Press E to launch the Starcade";
-        promptTarget = instance;
-        if (wasKeyJustPressed("KeyE")) {
-          audio.playEffect("dialogue");
-          openMiniGame();
-          notifyFirstSessionChecklist(CHECKLIST_EVENTS.MINI_GAME_LAUNCHED, {
-            interactable
-          });
-          showMessage(
-            {
-              text: "The arcade cabinet hums to life. Press Escape or Back to lobby to return.",
-              author: "Mission Log",
-              channel: "mission",
-              silent: true,
-              log: true
-            },
-            0
-          );
-        }
-      } else if (interactable.type === "fountain") {
-        promptText = "Press E to draw power from the fountain";
-        promptTarget = instance;
-        if (wasKeyJustPressed("KeyE")) {
-          if (interactable.charges > 0) {
-            interactable.charges -= 1;
-            playerStats.mp = clamp(playerStats.mp + 18, 0, playerStats.maxMp);
-            ui.refresh(playerStats);
-            audio.playEffect("fountain");
-            showMessage(
-              {
-                text: "Mana rush! Your MP was restored.",
-                author: "Mission Log",
-                channel: "mission"
-              },
-              3200
-            );
-          } else {
-            audio.playEffect("dialogue");
-            showMessage(
-              {
-                text: "The fountain needs time to recharge.",
-                author: "Mission Log",
-                channel: "mission"
-              },
-              3000
-            );
-          }
-        }
-      } else if (interactable.type === "npc") {
-        const guideName = getInteractableDisplayName(interactable);
-        promptText = `Press E to talk to ${guideName}`;
-        promptTarget = instance;
-        if (wasKeyJustPressed("KeyE")) {
-          notifyFirstSessionChecklist(CHECKLIST_EVENTS.NPC_INTERACTION, {
-            interactable,
-            guideName
-          });
-          const missionResult = completeMission(interactable.missionId);
-          audio.playEffect("dialogue");
-          if (missionResult.completed) {
-            const xpAward = missionResult.mission?.xp ?? 0;
-            const briefingLine = interactable.dialogue[0];
-            interactable.lineIndex = Math.min(1, interactable.dialogue.length - 1);
-            let message = `${briefingLine} +${xpAward} EXP.`;
-            if (missionResult.leveledUp) {
-              message += ` Level up! You reached level ${playerStats.level}.`;
-            }
-            showMessage(
-              { text: message, author: guideName, channel: "mission" },
-              5600
-            );
-          } else if (missionResult.locked) {
-            const required = Math.max(1, missionResult.mission?.requiredLevel ?? 1);
-            showMessage(
-              {
-                text: `Train to Level ${required} so ${guideName} can log your official briefing.`,
-                author: guideName,
-                channel: "mission"
-              },
-              4600
-            );
-          } else {
-            const line = interactable.dialogue[interactable.lineIndex];
-            interactable.lineIndex =
-              (interactable.lineIndex + 1) % interactable.dialogue.length;
-            showMessage(
-              { text: line, author: guideName, channel: "mission" },
-              4600
-            );
-          }
-        }
-      } else if (interactable.type === "comms") {
-        promptText = "Press E to access the comms console";
-        promptTarget = instance;
-        if (wasKeyJustPressed("KeyE")) {
-          const result = completeMission(interactable.missionId);
-          audio.playEffect("dialogue");
-          if (result.completed) {
-            const xpAward = result.mission?.xp ?? 0;
-            let message =
-              `You sync with the Astronaut account. Mission Control now follows your journey. +${xpAward} EXP.`;
-            if (result.leveledUp) {
-              message += ` Level up! You reached level ${playerStats.level}.`;
-            }
-            showMessage(
-              { text: message, author: "Mission Command", channel: "mission" },
-              5600
-            );
-          } else if (result.locked) {
-            const required = Math.max(1, result.mission?.requiredLevel ?? 1);
-            showMessage(
-              {
-                text: `Level ${required} clearance required to access the comms mission queue.`,
-                author: "Mission Command",
-                channel: "mission"
-              },
-              4400
-            );
-          } else if (result.alreadyComplete) {
-            showMessage(
-              {
-                text: "Mission Control feed already streams updates to your visor.",
-                author: "Mission Command",
-                channel: "mission"
-              },
-              4200
-            );
-          } else {
-            showMessage(
-              {
-                text: "The console hums, waiting for your next command.",
-                author: "Mission Command",
-                channel: "mission"
-              },
-              3600
-            );
-          }
-        }
-      }
-    }
-  }
-
-  let portalInstance = portal;
-  let nearPortal = false;
-  const portalOffsets = getWorldOffsetsAround(player.x);
-  for (const offset of portalOffsets) {
-    const candidate = createWorldInstance(portal, offset);
-    if (isNear(player, candidate, portal.interactionPadding)) {
-      portalInstance = candidate;
-      nearPortal = true;
-      break;
-    }
-  }
-
-  if (nearPortal) {
-    if (portalCharged) {
-      if (playerStats.level < portalRequiredLevel) {
-        promptText = `Reach Level ${portalRequiredLevel} to activate the portal.`;
-        promptTarget = portalInstance;
-        if (wasKeyJustPressed("KeyE")) {
-          audio.playEffect("dialogue");
-          showMessage(
-            {
-              text: `The portal resists you. Train until Level ${portalRequiredLevel} to stabilize the jump.`,
-              author: "Mission Command",
-              channel: "mission"
-            },
-            5200
-          );
-        }
-      } else if (portalCompleted) {
-        promptText = "The portal hums softly, its gateway already opened.";
-        promptTarget = portalInstance;
-      } else {
-        promptText = "Press E to step through the charged portal";
-        promptTarget = portalInstance;
-        if (wasKeyJustPressed("KeyE")) {
-          audio.playEffect("portalActivate");
-          queueEffect(
-            createPortalActivationEffect(
-              portalInstance.x + portalInstance.width / 2,
-              portalInstance.y + portalInstance.height / 2
-            )
-          );
-          queueEffect(
-            createScreenShakeEffect({ magnitude: 9, duration: 420, frequency: 12 })
-          );
-          portalCompleted = true;
-          const bonusExp = gainExperience(120);
-          playerStats.hp = playerStats.maxHp;
-          playerStats.mp = playerStats.maxMp;
-          ui.refresh(playerStats);
-          const highlightParts = [];
-          let completionMessage =
-            "You stride into the energized portal! +120 EXP. All stats restored for the journey ahead.";
-          if (bonusExp) {
-            highlightParts.push(`Level up! You reached level ${playerStats.level}.`);
-          }
-          const portalMission = completeMission("mission-portal-dive");
-          if (portalMission.completed) {
-            const xpAward = portalMission.mission?.xp ?? 0;
-            const title = portalMission.mission?.title ?? "Portal Mission";
-            highlightParts.push(`Mission Complete: ${title}! +${xpAward} EXP.`);
-            if (portalMission.leveledUp) {
-              highlightParts.push(`Level up! You reached level ${playerStats.level}.`);
-            }
-          } else if (portalMission.locked) {
-            const required = Math.max(1, portalMission.mission?.requiredLevel ?? 1);
-            highlightParts.push(`Reach Level ${required} to log the portal dive mission.`);
-          }
-          audio.playEffect("portalComplete");
-          for (const crystal of crystals) {
-            crystal.collected = false;
-          }
-          portalCharge = 0;
-          portalCharged = false;
-          crystalRunStartTime = 0;
-          ui.updateCrystals(portalCharge, crystals.length);
-          showMessage(
-            {
-              text:
-                highlightParts.length > 0
-                  ? `${completionMessage} ${highlightParts.join(" ")}`
-                  : completionMessage,
-              author: "Mission Log",
-              channel: "mission"
-            },
-            6200
-          );
-        }
-    }
-  } else {
-    promptText = "The portal is dormant. Gather more crystals.";
-    promptTarget = portalInstance;
-  }
-  }
-
-  if (movementHintVisible) {
-    if (ui && typeof ui.showMovementHint === "function") {
-      ui.showMovementHint(player, {
-        cameraOffset: cameraScrollX,
-        viewportWidth: viewport.width,
-        viewportHeight: viewport.height,
-        scale: renderScale,
-        onAcknowledge: acknowledgeMovementHint
-      });
-      ui.setPrompt("");
-    } else {
-      movementHintVisible = false;
-      ui.setPrompt(promptText, promptTarget);
-    }
-  } else {
-    ui.setPrompt(promptText, promptTarget);
-  }
-}
-
-function wrapOffset(value, length) {
-  if (!Number.isFinite(length) || length === 0) {
-    return 0;
-  }
-  let remainder = value % length;
-  if (remainder < 0) {
-    remainder += length;
-  }
-  return remainder;
-}
-
-function getWorldLoopOrigin(value) {
-  if (!Number.isFinite(worldWrapWidth) || worldWrapWidth <= 0) {
-    return 0;
-  }
-  const wrapped = wrapOffset(value, worldWrapWidth);
-  return value - wrapped;
-}
-
-function getWorldOffsetsAround(value) {
-  if (!Number.isFinite(worldWrapWidth) || worldWrapWidth <= 0) {
-    worldWrapOffsetsScratch.length = 1;
-    worldWrapOffsetsScratch[0] = 0;
-    return worldWrapOffsetsScratch;
-  }
-  const origin = getWorldLoopOrigin(value);
-  worldWrapOffsetsScratch.length = 3;
-  worldWrapOffsetsScratch[0] = origin - worldWrapWidth;
-  worldWrapOffsetsScratch[1] = origin;
-  worldWrapOffsetsScratch[2] = origin + worldWrapWidth;
-  return worldWrapOffsetsScratch;
-}
-
-function getWorldWrapOffsetsForView(cameraOffset, margin = 0) {
-  if (!Number.isFinite(worldWrapWidth) || worldWrapWidth <= 0) {
-    worldWrapRenderOffsetsScratch.length = 1;
-    worldWrapRenderOffsetsScratch[0] = 0;
-    return worldWrapRenderOffsetsScratch;
-  }
-  const cameraLeft = cameraOffset - margin;
-  const cameraRight = cameraOffset + viewport.width + margin;
-  const startIndex = Math.floor(cameraLeft / worldWrapWidth);
-  const endIndex = Math.floor(cameraRight / worldWrapWidth);
-  worldWrapRenderOffsetsScratch.length = 0;
-  for (let index = startIndex; index <= endIndex; index += 1) {
-    worldWrapRenderOffsetsScratch.push(index * worldWrapWidth);
-  }
-  if (worldWrapRenderOffsetsScratch.length === 0) {
-    worldWrapRenderOffsetsScratch.push(0);
-  }
-  return worldWrapRenderOffsetsScratch;
-}
+);
 
 function createLobbyLayoutEditor(options = {}) {
   const {
@@ -6910,1161 +6396,6 @@ function isWorldInstanceVisible(x, width, cameraOffset, margin = 96) {
   const left = cameraOffset - margin;
   const right = cameraOffset + viewport.width + margin;
   return x + width > left && x < right;
-}
-
-function drawParallaxBackgroundImage(ctx) {
-  let drewLayer = false;
-
-  for (const layer of parallaxLayers) {
-    if (!layer.ready) {
-      continue;
-    }
-
-    const { width: sourceWidth, height: sourceHeight } = layer.dimensions;
-    if (!(sourceWidth > 0) || !(sourceHeight > 0)) {
-      continue;
-    }
-
-    const widthScale = viewport.width / sourceWidth;
-    const heightScale = viewport.height / sourceHeight;
-    const scale = Math.min(widthScale, heightScale);
-    const drawWidth = sourceWidth * scale;
-    const drawHeight = sourceHeight * scale;
-
-    if (!(drawWidth > 0) || !(drawHeight > 0)) {
-      continue;
-    }
-
-    let offsetY;
-    if (layer.align === "bottom") {
-      offsetY = viewport.height - drawHeight;
-    } else if (layer.align === "top") {
-      offsetY = 0;
-    } else {
-      offsetY = (viewport.height - drawHeight) / 2;
-    }
-    if (typeof layer.offsetY === "number" && Number.isFinite(layer.offsetY)) {
-      offsetY += layer.offsetY;
-    }
-
-    const parallaxFactor = Number.isFinite(layer.speed) ? layer.speed : 0;
-    const scrollOffset = wrapOffset(parallaxScroll * parallaxFactor, drawWidth);
-
-    ctx.save();
-    if (typeof layer.opacity === "number" && layer.opacity >= 0) {
-      ctx.globalAlpha = Math.max(0, Math.min(1, layer.opacity));
-    }
-
-    for (let x = -drawWidth; x <= viewport.width + drawWidth; x += drawWidth) {
-      const drawX = Math.round(x - scrollOffset);
-      ctx.drawImage(
-        layer.image,
-        0,
-        0,
-        sourceWidth,
-        sourceHeight,
-        drawX,
-        offsetY,
-        drawWidth,
-        drawHeight
-      );
-    }
-
-    ctx.restore();
-    drewLayer = true;
-  }
-
-  if (!drewLayer) {
-    return;
-  }
-
-  const haze = ctx.createLinearGradient(0, 0, 0, viewport.height);
-  haze.addColorStop(0, "rgba(9, 15, 32, 0)");
-  haze.addColorStop(0.55, "rgba(9, 15, 32, 0.18)");
-  haze.addColorStop(1, "rgba(9, 15, 32, 0.55)");
-  ctx.fillStyle = haze;
-  ctx.fillRect(0, 0, viewport.width, viewport.height);
-}
-
-function drawGroundPlane(ctx) {
-  const groundHeight = viewport.height - groundY;
-  const baseGradient = ctx.createLinearGradient(0, groundY, 0, viewport.height);
-  baseGradient.addColorStop(0, "#16312a");
-  baseGradient.addColorStop(1, "#061612");
-  ctx.fillStyle = baseGradient;
-  ctx.fillRect(0, groundY, viewport.width, groundHeight);
-
-  const tileWidth = 76;
-  const offset = wrapOffset(parallaxScroll * 1.24, tileWidth);
-  for (let x = -tileWidth; x <= viewport.width + tileWidth; x += tileWidth) {
-    const drawX = x - offset;
-    const stripeGradient = ctx.createLinearGradient(drawX, groundY, drawX, groundY + groundHeight);
-    stripeGradient.addColorStop(0, "rgba(124, 243, 216, 0.18)");
-    stripeGradient.addColorStop(0.7, "rgba(30, 66, 52, 0.12)");
-    stripeGradient.addColorStop(1, "rgba(10, 24, 20, 0)");
-    ctx.fillStyle = stripeGradient;
-    ctx.fillRect(drawX, groundY, tileWidth, groundHeight);
-
-    const rimGradient = ctx.createLinearGradient(drawX, groundY - 6, drawX, groundY + 18);
-    rimGradient.addColorStop(0, "rgba(200, 255, 244, 0.35)");
-    rimGradient.addColorStop(1, "rgba(28, 64, 46, 0)");
-    ctx.fillStyle = rimGradient;
-    ctx.fillRect(drawX, groundY - 6, tileWidth, 24);
-  }
-}
-
-function renderParallaxBackdrop(ctx, time) {
-  ctx.fillStyle = getFallbackBackgroundGradient();
-  ctx.fillRect(0, 0, viewport.width, viewport.height);
-
-  drawParallaxBackgroundImage(ctx);
-  drawGroundPlane(ctx);
-}
-
-const activeEffects = [];
-
-function getCurrentTime() {
-  if (typeof performance !== "undefined" && typeof performance.now === "function") {
-    return performance.now();
-  }
-  return Date.now();
-}
-
-function queueEffect(effect) {
-  if (!effect || typeof effect !== "object") {
-    return;
-  }
-  if (typeof effect.startedAt !== "number") {
-    effect.startedAt = getCurrentTime();
-  }
-  if (typeof effect.duration !== "number" || effect.duration <= 0) {
-    return;
-  }
-  activeEffects.push(effect);
-}
-
-function createParticleBurstEffect(originX, originY, options = {}) {
-  const now = getCurrentTime();
-  const count = Math.max(1, Math.floor(options.count ?? 12));
-  const duration = Math.max(16, options.duration ?? 480);
-  const palette = Array.isArray(options.palette) && options.palette.length > 0
-    ? options.palette
-    : ["rgba(220, 255, 244, 0.85)"];
-  const baseAngle = typeof options.baseAngle === "number" ? options.baseAngle : -Math.PI / 2;
-  const spread = typeof options.spread === "number" ? options.spread : Math.PI * 2;
-  const speedMin = typeof options.speedMin === "number" ? options.speedMin : 60;
-  const speedMax = typeof options.speedMax === "number" ? options.speedMax : 160;
-  const sizeMin = typeof options.sizeMin === "number" ? options.sizeMin : 1.5;
-  const sizeMax = typeof options.sizeMax === "number" ? options.sizeMax : 3.5;
-  const gravity = typeof options.gravity === "number" ? options.gravity : 0;
-  const fade = options.fade !== false;
-
-  const particles = Array.from({ length: count }, () => {
-    const angle = baseAngle + (Math.random() - 0.5) * spread;
-    const speed = speedMin + Math.random() * Math.max(0, speedMax - speedMin);
-    const size = sizeMin + Math.random() * Math.max(0, sizeMax - sizeMin);
-    const color = palette[Math.floor(Math.random() * palette.length)] ?? palette[0];
-    return {
-      vx: Math.cos(angle) * speed,
-      vy: Math.sin(angle) * speed,
-      life: duration,
-      size,
-      color
-    };
-  });
-
-  return {
-    type: "particleBurst",
-    originX,
-    originY,
-    duration,
-    startedAt: now,
-    gravity,
-    fade,
-    particles
-  };
-}
-
-function createScreenShakeEffect(options = {}) {
-  const now = getCurrentTime();
-  const duration = Math.max(16, options.duration ?? 300);
-  const magnitude = typeof options.magnitude === "number" ? options.magnitude : 6;
-  const frequency = typeof options.frequency === "number" ? options.frequency : 11;
-  return {
-    type: "screenShake",
-    duration,
-    startedAt: now,
-    magnitude,
-    frequency,
-    seedX: Math.random() * Math.PI * 2,
-    seedY: Math.random() * Math.PI * 2
-  };
-}
-
-function createJumpDustEffect(x, y) {
-  return createParticleBurstEffect(x, y, {
-    count: 10,
-    duration: 360,
-    palette: ["rgba(120, 230, 210, 0.5)", "rgba(190, 255, 240, 0.65)"],
-    baseAngle: -Math.PI / 2,
-    spread: Math.PI * 0.75,
-    speedMin: 80,
-    speedMax: 180,
-    gravity: 420,
-    sizeMin: 1.5,
-    sizeMax: 3.2
-  });
-}
-
-function createCrystalBurstEffect(x, y) {
-  return createParticleBurstEffect(x, y, {
-    count: 14,
-    duration: 520,
-    palette: ["rgba(120, 255, 230, 0.9)", "rgba(180, 255, 250, 0.95)", "rgba(255, 255, 255, 0.8)"],
-    baseAngle: -Math.PI / 2,
-    spread: Math.PI * 1.5,
-    speedMin: 90,
-    speedMax: 200,
-    gravity: 60,
-    sizeMin: 1.2,
-    sizeMax: 2.8
-  });
-}
-
-function createPortalActivationEffect(x, y) {
-  return createParticleBurstEffect(x, y, {
-    count: 26,
-    duration: 640,
-    palette: [
-      "rgba(130, 245, 255, 0.95)",
-      "rgba(90, 170, 255, 0.85)",
-      "rgba(200, 255, 255, 0.75)"
-    ],
-    baseAngle: -Math.PI / 2,
-    spread: Math.PI * 2,
-    speedMin: 120,
-    speedMax: 260,
-    gravity: 120,
-    sizeMin: 1.8,
-    sizeMax: 3.8
-  });
-}
-
-function getScreenShakeOffset(now) {
-  let offsetX = 0;
-  let offsetY = 0;
-  for (const effect of activeEffects) {
-    if (effect.type !== "screenShake") {
-      continue;
-    }
-    const elapsed = now - effect.startedAt;
-    if (elapsed < 0 || elapsed >= effect.duration) {
-      continue;
-    }
-    const decay = 1 - elapsed / effect.duration;
-    const frequency = effect.frequency ?? 11;
-    const timeSeconds = elapsed / 1000;
-    offsetX += Math.sin(timeSeconds * frequency * Math.PI * 2 + effect.seedX) * effect.magnitude * decay;
-    offsetY += Math.cos(timeSeconds * (frequency * 0.8) * Math.PI * 2 + effect.seedY) * effect.magnitude * 0.6 * decay;
-  }
-  return { offsetX, offsetY };
-}
-
-function renderWorldEffects(ctx, now) {
-  if (!activeEffects.length) {
-    return;
-  }
-  for (const effect of activeEffects) {
-    if (effect.type !== "particleBurst") {
-      continue;
-    }
-    const elapsed = now - effect.startedAt;
-    if (elapsed < 0 || elapsed >= effect.duration) {
-      continue;
-    }
-    ctx.save();
-    for (const particle of effect.particles) {
-      const particleAge = Math.min(elapsed, particle.life);
-      const seconds = particleAge / 1000;
-      const fade = effect.fade ? 1 - particleAge / particle.life : 1;
-      const alpha = Math.max(0, Math.min(1, fade));
-      if (alpha <= 0) {
-        continue;
-      }
-      const px = effect.originX + particle.vx * seconds;
-      const py = effect.originY + particle.vy * seconds + (effect.gravity * seconds * seconds) / 2;
-      ctx.globalAlpha = alpha;
-      ctx.fillStyle = particle.color;
-      ctx.beginPath();
-      ctx.arc(px, py, particle.size, 0, Math.PI * 2);
-      ctx.fill();
-    }
-    ctx.restore();
-  }
-}
-
-function cleanupExpiredEffects(now) {
-  if (!activeEffects.length) {
-    return;
-  }
-  for (let index = activeEffects.length - 1; index >= 0; index -= 1) {
-    const effect = activeEffects[index];
-    const elapsed = now - effect.startedAt;
-    if (elapsed >= effect.duration) {
-      activeEffects.splice(index, 1);
-    }
-  }
-}
-
-function render(timestamp) {
-  const now =
-    typeof performance !== "undefined" && typeof performance.now === "function"
-      ? timestamp
-      : Date.now();
-  const time = timestamp / 1000;
-  const shakeOffset = getScreenShakeOffset(now);
-
-  ctx.setTransform(
-    renderScale * devicePixelScale,
-    0,
-    0,
-    renderScale * devicePixelScale,
-    0,
-    0
-  );
-
-  renderParallaxBackdrop(ctx, time);
-
-  ctx.save();
-  ctx.translate(-cameraScrollX + shakeOffset.offsetX, shakeOffset.offsetY);
-
-  const renderOffsets = getWorldWrapOffsetsForView(cameraScrollX, viewport.width);
-
-  if (platformSprite.isReady() && platformSprite.image) {
-    ctx.save();
-    ctx.imageSmoothingEnabled = false;
-    for (const offset of renderOffsets) {
-      for (const platform of platforms) {
-        const platformX = platform.x + offset;
-        if (!isWorldInstanceVisible(platformX, platform.width, cameraScrollX)) {
-          continue;
-        }
-        ctx.drawImage(
-          platformSprite.image,
-          0,
-          0,
-          platformSprite.image.width,
-          platformSprite.image.height,
-          platformX,
-          platform.y,
-          platform.width,
-          platform.height
-        );
-      }
-    }
-    ctx.restore();
-  } else {
-    ctx.fillStyle = "#3b5e3f";
-    for (const offset of renderOffsets) {
-      for (const platform of platforms) {
-        const platformX = platform.x + offset;
-        if (!isWorldInstanceVisible(platformX, platform.width, cameraScrollX)) {
-          continue;
-        }
-        drawRoundedRect(platformX, platform.y, platform.width, platform.height, 6);
-      }
-    }
-  }
-
-  for (const offset of renderOffsets) {
-    const portalInstance = createWorldInstance(portal, offset);
-    if (!isWorldInstanceVisible(portalInstance.x, portalInstance.width, cameraScrollX)) {
-      continue;
-    }
-    drawPortal(portalInstance, time);
-  }
-
-  for (const offset of renderOffsets) {
-    for (const interactable of interactables) {
-      const instance = createWorldInstance(interactable, offset);
-      const width = instance.width ?? 0;
-      if (!isWorldInstanceVisible(instance.x, width, cameraScrollX, 64)) {
-        continue;
-      }
-      if (interactable.type === "bulletin") {
-        drawBulletin(instance, time);
-      } else if (interactable.type === "chest") {
-        drawChest(instance);
-      } else if (interactable.type === "arcade") {
-        drawArcade(instance, time);
-      } else if (interactable.type === "fountain") {
-        drawFountain(instance, time);
-      } else if (interactable.type === "npc") {
-        drawGuide(instance, time);
-      } else if (interactable.type === "comms") {
-        drawComms(instance, time);
-      }
-    }
-  }
-
-  for (const offset of renderOffsets) {
-    for (const crystal of crystals) {
-      if (crystal.collected) continue;
-      const instance = createWorldInstance(crystal, offset);
-      if (!isWorldInstanceVisible(instance.x - instance.radius, instance.radius * 2, cameraScrollX, 64)) {
-        continue;
-      }
-      drawCrystal(instance, time);
-    }
-  }
-
-  drawPlayer(player, time);
-
-  renderWorldEffects(ctx, now);
-
-  ctx.restore();
-
-  if (layoutEditor && typeof layoutEditor.drawOverlay === "function" && layoutEditor.isActive()) {
-    layoutEditor.drawOverlay(ctx);
-  }
-
-  if (ui.promptText) {
-    drawPromptBubble(ui.promptText, ui.promptEntity || player);
-  }
-
-  cleanupExpiredEffects(now);
-}
-
-function drawPortal(portalInstance, time) {
-  const { x: portalX, y: portalY, width: portalWidth, height: portalHeight } = portalInstance;
-  const framePulse = (Math.sin(time * (portalCharged ? 4.4 : 2.2)) + 1) / 2;
-  const archColor = portalCharged ? "#4a2f7f" : "#384d6b";
-  const glowInner = portalCharged ? "rgba(160, 255, 245, 0.95)" : "rgba(150, 205, 255, 0.65)";
-  const glowOuter = portalCharged ? "rgba(90, 220, 200, 0.35)" : "rgba(100, 140, 220, 0.1)";
-
-  ctx.save();
-  ctx.fillStyle = archColor;
-  drawRoundedRect(portalX - 12, portalY - 12, portalWidth + 24, portalHeight + 24, 24);
-
-  const glowGradient = ctx.createRadialGradient(
-    portalX + portalWidth / 2,
-    portalY + portalHeight / 2,
-    10,
-    portalX + portalWidth / 2,
-    portalY + portalHeight / 2,
-    portalWidth / 2
-  );
-  glowGradient.addColorStop(0, glowInner);
-  glowGradient.addColorStop(1, glowOuter);
-  ctx.fillStyle = glowGradient;
-  ctx.globalAlpha = portalCharged ? 1 : 0.85;
-  drawRoundedRect(portalX, portalY, portalWidth, portalHeight, 20);
-
-  ctx.globalAlpha = 1;
-  ctx.strokeStyle = portalCharged
-    ? `rgba(230, 250, 255, ${0.55 + framePulse * 0.25})`
-    : `rgba(200, 240, 255, 0.55)`;
-  ctx.lineWidth = 4;
-  ctx.strokeRect(portalX + 12, portalY + 12, portalWidth - 24, portalHeight - 24);
-
-  const pulse = (Math.sin(time * (portalCharged ? 5.2 : 2.4)) + 1) / 2;
-  const ellipseColor = portalCharged
-    ? `rgba(120, 255, 225, ${0.5 + pulse * 0.4})`
-    : `rgba(180, 230, 255, ${0.35 + pulse * 0.35})`;
-  ctx.lineWidth = portalCharged ? 4 : 3;
-  ctx.strokeStyle = ellipseColor;
-  ctx.beginPath();
-  ctx.ellipse(
-    portalX + portalWidth / 2,
-    portalY + portalHeight / 2,
-    24 + pulse * (portalCharged ? 14 : 8),
-    60 + pulse * (portalCharged ? 20 : 12),
-    0,
-    0,
-    Math.PI * 2
-  );
-  ctx.stroke();
-
-  if (portalCharged) {
-    ctx.fillStyle = "rgba(255, 255, 255, 0.6)";
-    for (let i = 0; i < 6; i += 1) {
-      const angle = time * 1.6 + (i * Math.PI) / 3;
-      const radius = 32 + framePulse * 12;
-      const orbX = portalX + portalWidth / 2 + Math.cos(angle) * radius;
-      const orbY = portalY + portalHeight / 2 + Math.sin(angle) * (radius * 0.6);
-      const orbSize = 6 + Math.sin(time * 4 + i) * 2;
-      ctx.beginPath();
-      ctx.arc(orbX, orbY, orbSize, 0, Math.PI * 2);
-      ctx.fill();
-    }
-  }
-  ctx.restore();
-}
-
-function drawPlayer(entity, time) {
-  ctx.save();
-  ctx.translate(entity.x + entity.width / 2, entity.y + entity.height);
-  ctx.scale(entity.direction, 1);
-  ctx.translate(-entity.width / 2, -entity.height);
-
-  const appearance = entity.appearance ?? playerAppearance;
-
-  if (playerSpriteState.isReady()) {
-    const spriteImage = playerSpriteState.image;
-    ctx.imageSmoothingEnabled = false;
-    ctx.drawImage(
-      spriteImage,
-      0,
-      0,
-      spriteImage.width,
-      spriteImage.height,
-      0,
-      0,
-      entity.width,
-      entity.height
-    );
-  } else {
-    ctx.fillStyle = appearance.shirt;
-    drawRoundedRect(4, 12, entity.width - 8, entity.height - 18, 6);
-
-    ctx.fillStyle = appearance.skin;
-    drawRoundedRect(6, 0, entity.width - 12, 18, 6);
-
-    ctx.fillStyle = appearance.hair;
-    ctx.fillRect(8, 16, entity.width - 16, 6);
-
-    ctx.fillStyle = "#111";
-    const blink = Math.sin(time * 3.2) > -0.2 ? 1 : 0.2;
-    ctx.fillRect(12, 6, 4, 4 * blink);
-    ctx.fillRect(entity.width - 16, 6, 4, 4 * blink);
-
-    ctx.fillStyle = "#c9d7ff";
-    ctx.fillRect(0, entity.height - 12, entity.width - 12, 12);
-    ctx.fillRect(entity.width - 12, entity.height - 8, 12, 8);
-  }
-  ctx.restore();
-}
-
-function drawPromptBubble(text, entity) {
-  if (!text) {
-    return;
-  }
-
-  const target = entity ?? player;
-  const targetWidth = typeof target.width === "number" ? target.width : 0;
-  const targetX = (typeof target.x === "number" ? target.x : 0) - cameraScrollX;
-  const centerX = targetX + targetWidth / 2;
-  ctx.save();
-  ctx.font = promptFont;
-  ctx.textBaseline = "alphabetic";
-  ctx.textAlign = "left";
-
-  const metrics = getPromptMetrics(text);
-  const paddingX = 18;
-  const paddingY = 14;
-  const tailHeight = 18;
-  const tailHalfWidth = 20;
-  const radius = 24;
-  const bubbleWidth = metrics.width + paddingX * 2;
-  const minimumHeight = 48;
-  const bubbleHeight = Math.max(metrics.height + paddingY * 2, minimumHeight);
-  const anchorTop =
-    typeof target.promptAnchorY === "number" ? target.promptAnchorY : target.y ?? 0;
-  const anchorHeight = target.height ?? 0;
-  const tailGap = Math.min(14, Math.max(6, anchorHeight ? anchorHeight * 0.12 : 8));
-  const marginX = 18;
-  let bubbleX = centerX - bubbleWidth / 2;
-  bubbleX = Math.max(marginX, Math.min(bubbleX, viewport.width - bubbleWidth - marginX));
-  let bubbleY = anchorTop - bubbleHeight - tailHeight - tailGap;
-  bubbleY = Math.max(18, bubbleY);
-  const tailBaseY = bubbleY + bubbleHeight;
-  const tailTipX = Math.max(
-    bubbleX + radius + 6,
-    Math.min(centerX, bubbleX + bubbleWidth - radius - 6)
-  );
-  const tailTipY = Math.max(
-    tailBaseY + 6,
-    Math.min(tailBaseY + tailHeight, anchorTop - Math.max(2, tailGap * 0.5))
-  );
-
-  const traceBubblePath = () => {
-    ctx.beginPath();
-    ctx.moveTo(bubbleX + radius, bubbleY);
-    ctx.lineTo(bubbleX + bubbleWidth - radius, bubbleY);
-    ctx.quadraticCurveTo(
-      bubbleX + bubbleWidth,
-      bubbleY,
-      bubbleX + bubbleWidth,
-      bubbleY + radius
-    );
-    ctx.lineTo(bubbleX + bubbleWidth, tailBaseY - radius);
-    ctx.quadraticCurveTo(
-      bubbleX + bubbleWidth,
-      tailBaseY,
-      bubbleX + bubbleWidth - radius,
-      tailBaseY
-    );
-    ctx.lineTo(tailTipX + tailHalfWidth, tailBaseY);
-    ctx.lineTo(tailTipX, tailTipY);
-    ctx.lineTo(tailTipX - tailHalfWidth, tailBaseY);
-    ctx.lineTo(bubbleX + radius, tailBaseY);
-    ctx.quadraticCurveTo(bubbleX, tailBaseY, bubbleX, tailBaseY - radius);
-    ctx.lineTo(bubbleX, bubbleY + radius);
-    ctx.quadraticCurveTo(bubbleX, bubbleY, bubbleX + radius, bubbleY);
-    ctx.closePath();
-  };
-
-  ctx.save();
-  ctx.shadowColor = "rgba(35, 20, 68, 0.35)";
-  ctx.shadowBlur = 18;
-  ctx.shadowOffsetX = 0;
-  ctx.shadowOffsetY = 12;
-  traceBubblePath();
-  const fillGradient = ctx.createLinearGradient(
-    bubbleX,
-    bubbleY,
-    bubbleX,
-    tailBaseY
-  );
-  fillGradient.addColorStop(0, "rgba(255, 255, 255, 0.96)");
-  fillGradient.addColorStop(1, "rgba(255, 223, 246, 0.96)");
-  ctx.fillStyle = fillGradient;
-  ctx.fill();
-  ctx.restore();
-
-  ctx.save();
-  traceBubblePath();
-  ctx.lineWidth = 4;
-  ctx.strokeStyle = "#2f47ff";
-  ctx.stroke();
-  ctx.restore();
-
-  const innerRadius = Math.max(12, radius - 6);
-  const innerX = bubbleX + 14;
-  const innerY = bubbleY + 12;
-  const innerWidth = Math.max(0, bubbleWidth - 28);
-  const innerHeight = Math.max(0, bubbleHeight - 24);
-
-  ctx.save();
-  ctx.beginPath();
-  ctx.moveTo(innerX + innerRadius, innerY);
-  ctx.lineTo(innerX + innerWidth - innerRadius, innerY);
-  ctx.quadraticCurveTo(
-    innerX + innerWidth,
-    innerY,
-    innerX + innerWidth,
-    innerY + innerRadius
-  );
-  ctx.lineTo(innerX + innerWidth, innerY + innerHeight - innerRadius);
-  ctx.quadraticCurveTo(
-    innerX + innerWidth,
-    innerY + innerHeight,
-    innerX + innerWidth - innerRadius,
-    innerY + innerHeight
-  );
-  ctx.lineTo(innerX + innerRadius, innerY + innerHeight);
-  ctx.quadraticCurveTo(innerX, innerY + innerHeight, innerX, innerY + innerHeight - innerRadius);
-  ctx.lineTo(innerX, innerY + innerRadius);
-  ctx.quadraticCurveTo(innerX, innerY, innerX + innerRadius, innerY);
-  ctx.closePath();
-  ctx.setLineDash([8, 10]);
-  ctx.lineDashOffset = 4;
-  ctx.lineWidth = 2;
-  ctx.strokeStyle = "rgba(255, 188, 116, 0.55)";
-  ctx.stroke();
-  ctx.restore();
-
-  ctx.save();
-  const highlightGradient = ctx.createLinearGradient(
-    bubbleX,
-    bubbleY,
-    bubbleX,
-    bubbleY + bubbleHeight * 0.6
-  );
-  highlightGradient.addColorStop(0, "rgba(255, 255, 255, 0.65)");
-  highlightGradient.addColorStop(1, "rgba(255, 255, 255, 0)");
-  traceBubblePath();
-  ctx.clip();
-  ctx.fillStyle = highlightGradient;
-  ctx.fillRect(bubbleX, bubbleY, bubbleWidth, bubbleHeight);
-  ctx.restore();
-
-  const ascent = metrics.ascent || metrics.height * 0.75;
-  const textX = bubbleX + (bubbleWidth - metrics.width) / 2;
-  const textY = bubbleY + paddingY + ascent;
-  ctx.fillStyle = "#2a1d4f";
-  ctx.shadowColor = "rgba(255, 206, 115, 0.7)";
-  ctx.shadowBlur = 0;
-  ctx.shadowOffsetX = 0;
-  ctx.shadowOffsetY = 2;
-  ctx.fillText(text, textX, textY);
-  ctx.restore();
-
-  ctx.restore();
-}
-
-function drawBulletin(board, time) {
-  ctx.save();
-  ctx.translate(board.x, board.y);
-  const pulse = (Math.sin(time * 2.2) + 1) / 2;
-  ctx.fillStyle = "#1d253f";
-  drawRoundedRect(0, 10, board.width, board.height - 10, 10);
-  ctx.fillStyle = "#151c2c";
-  ctx.fillRect(6, board.height - 12, 12, 12);
-  ctx.fillRect(board.width - 18, board.height - 12, 12, 12);
-  ctx.fillStyle = `rgba(120, 150, 255, ${0.35 + pulse * 0.25})`;
-  drawRoundedRect(8, 16, board.width - 16, board.height - 38, 8);
-  ctx.fillStyle = "rgba(15, 20, 40, 0.85)";
-  drawRoundedRect(14, 24, board.width - 28, 32, 6);
-  ctx.fillStyle = "rgba(215, 230, 255, 0.82)";
-  ctx.fillRect(20, 30, board.width - 40, 6);
-  ctx.fillRect(20, 40, board.width - 56, 4);
-  ctx.fillStyle = `rgba(255, 200, 255, ${0.3 + pulse * 0.3})`;
-  ctx.fillRect(20, 48, board.width - 40, 4);
-  ctx.fillStyle = `rgba(140, 200, 255, ${0.35 + pulse * 0.35})`;
-  drawRoundedRect(14, board.height - 28, board.width - 28, 10, 4);
-  ctx.restore();
-}
-
-function drawChest(chest) {
-  ctx.save();
-  ctx.translate(chest.x, chest.y);
-  if (chestSprite.isReady() && chestSprite.image) {
-    ctx.imageSmoothingEnabled = false;
-    ctx.drawImage(
-      chestSprite.image,
-      0,
-      0,
-      chestSprite.image.width,
-      chestSprite.image.height,
-      0,
-      0,
-      chest.width,
-      chest.height
-    );
-    ctx.restore();
-    return;
-  }
-  ctx.fillStyle = chest.opened ? "#a77b3b" : "#c58f3d";
-  drawRoundedRect(0, 10, chest.width, chest.height - 10, 6);
-  ctx.fillStyle = chest.opened ? "#8a5f23" : "#a16b22";
-  drawRoundedRect(0, 0, chest.width, 18, 6);
-  ctx.fillStyle = "#f7d774";
-  ctx.fillRect(chest.width / 2 - 4, 18, 8, 10);
-  ctx.restore();
-}
-
-function drawArcade(cabinet, time) {
-  ctx.save();
-  ctx.translate(cabinet.x, cabinet.y);
-  if (arcadeSprite.isReady() && arcadeSprite.image) {
-    ctx.imageSmoothingEnabled = false;
-    ctx.drawImage(
-      arcadeSprite.image,
-      0,
-      0,
-      arcadeSprite.image.width,
-      arcadeSprite.image.height,
-      0,
-      0,
-      cabinet.width,
-      cabinet.height
-    );
-    ctx.restore();
-    return;
-  }
-
-  const glow = (Math.sin(time * 3.1) + 1) / 2;
-  ctx.fillStyle = "#1a1d33";
-  drawRoundedRect(0, 8, cabinet.width, cabinet.height - 8, 10);
-  ctx.fillStyle = "#2c3563";
-  drawRoundedRect(4, 16, cabinet.width - 8, cabinet.height - 24, 8);
-  const screenHeight = Math.min(56, cabinet.height * 0.48);
-  ctx.fillStyle = `rgba(110, 205, 255, ${0.3 + glow * 0.45})`;
-  drawRoundedRect(12, 24, cabinet.width - 24, screenHeight, 6);
-  ctx.fillStyle = "#040713";
-  drawRoundedRect(14, 26, cabinet.width - 28, screenHeight - 8, 5);
-  ctx.fillStyle = `rgba(255, 188, 255, ${0.4 + glow * 0.35})`;
-  drawRoundedRect(cabinet.width / 2 - 18, cabinet.height - 48, 36, 18, 6);
-  ctx.fillStyle = `rgba(255, 214, 126, ${0.7 + glow * 0.2})`;
-  ctx.beginPath();
-  ctx.arc(cabinet.width / 2 - 26, cabinet.height - 34, 6, 0, Math.PI * 2);
-  ctx.arc(cabinet.width / 2 + 26, cabinet.height - 34, 6, 0, Math.PI * 2);
-  ctx.fill();
-  ctx.restore();
-}
-
-function drawFountain(fountain, time) {
-  ctx.save();
-  ctx.translate(fountain.x, fountain.y);
-  if (fountainSprite.isReady() && fountainSprite.image) {
-    ctx.imageSmoothingEnabled = false;
-    ctx.drawImage(
-      fountainSprite.image,
-      0,
-      0,
-      fountainSprite.image.width,
-      fountainSprite.image.height,
-      0,
-      0,
-      fountain.width,
-      fountain.height
-    );
-    ctx.restore();
-    return;
-  }
-  ctx.fillStyle = "#3c4a62";
-  drawRoundedRect(0, fountain.height - 18, fountain.width, 18, 8);
-  ctx.fillStyle = "#556b8f";
-  drawRoundedRect(6, 14, fountain.width - 12, fountain.height - 32, 10);
-  const pulse = (Math.sin(time * 2) + 1) / 2;
-  ctx.fillStyle = `rgba(120, 205, 255, ${0.4 + pulse * 0.4})`;
-  drawRoundedRect(10, 20, fountain.width - 20, fountain.height - 40, 10);
-  ctx.restore();
-}
-
-function drawGuide(guide, time) {
-  ctx.save();
-  ctx.translate(guide.x, guide.y);
-  if (guideSprite.isReady() && guideSprite.image) {
-    ctx.imageSmoothingEnabled = false;
-    ctx.drawImage(
-      guideSprite.image,
-      0,
-      0,
-      guideSprite.image.width,
-      guideSprite.image.height,
-      0,
-      0,
-      guide.width,
-      guide.height
-    );
-    ctx.restore();
-    return;
-  }
-  ctx.fillStyle = "#f4dede";
-  drawRoundedRect(4, 8, guide.width - 8, guide.height - 12, 10);
-  ctx.fillStyle = "#dba6ff";
-  drawRoundedRect(8, guide.height - 28, guide.width - 16, 20, 8);
-  ctx.fillStyle = "#000";
-  const bob = Math.sin(time * 2.4) * 1.5;
-  ctx.fillRect(12, 14 + bob, 4, 6);
-  ctx.fillRect(guide.width - 16, 14 + bob, 4, 6);
-  ctx.fillStyle = "#fff";
-  ctx.fillRect(12, 14 + bob, 4, 4);
-  ctx.fillRect(guide.width - 16, 14 + bob, 4, 4);
-  ctx.fillStyle = "#fefefe";
-  ctx.beginPath();
-  ctx.arc(guide.width / 2, 8, 10, Math.PI, Math.PI * 2);
-  ctx.fill();
-  ctx.restore();
-}
-
-function drawComms(consoleUnit, time) {
-  ctx.save();
-  ctx.translate(consoleUnit.x, consoleUnit.y);
-  const pulse = (Math.sin(time * 3.4) + 1) / 2;
-  ctx.fillStyle = "#1c273a";
-  drawRoundedRect(0, 12, consoleUnit.width, consoleUnit.height - 12, 12);
-  ctx.fillStyle = "#2f3f62";
-  drawRoundedRect(6, 0, consoleUnit.width - 12, 36, 10);
-  ctx.fillStyle = `rgba(110, 215, 255, ${0.45 + pulse * 0.35})`;
-  drawRoundedRect(12, 8, consoleUnit.width - 24, 22, 8);
-  ctx.strokeStyle = `rgba(140, 200, 255, ${0.6 + pulse * 0.2})`;
-  ctx.lineWidth = 3;
-  ctx.beginPath();
-  ctx.moveTo(consoleUnit.width / 2, -6);
-  ctx.lineTo(consoleUnit.width / 2, 6);
-  ctx.stroke();
-  ctx.beginPath();
-  ctx.arc(consoleUnit.width / 2, -8, 6 + pulse * 3, 0, Math.PI * 2);
-  ctx.stroke();
-  ctx.fillStyle = `rgba(90, 255, 200, ${0.3 + pulse * 0.4})`;
-  drawRoundedRect(10, consoleUnit.height - 20, consoleUnit.width - 20, 10, 4);
-  ctx.restore();
-}
-
-function drawCrystal(crystal, time) {
-  ctx.save();
-  ctx.translate(crystal.x, crystal.y);
-  if (crystalSprite.isReady() && crystalSprite.image) {
-    const targetSize = crystal.radius * 2 + 12;
-    ctx.imageSmoothingEnabled = false;
-    ctx.drawImage(
-      crystalSprite.image,
-      0,
-      0,
-      crystalSprite.image.width,
-      crystalSprite.image.height,
-      -targetSize / 2,
-      -targetSize / 2,
-      targetSize,
-      targetSize
-    );
-    ctx.restore();
-    return;
-  }
-  ctx.rotate(Math.sin(time * 2 + crystal.x * 0.01) * 0.1);
-  const gradient = ctx.createLinearGradient(-crystal.radius, -crystal.radius, crystal.radius, crystal.radius);
-  gradient.addColorStop(0, "#d9baff");
-  gradient.addColorStop(1, "#8fb5ff");
-  ctx.fillStyle = gradient;
-  ctx.beginPath();
-  ctx.moveTo(0, -crystal.radius - 6);
-  ctx.lineTo(crystal.radius, 0);
-  ctx.lineTo(0, crystal.radius + 6);
-  ctx.lineTo(-crystal.radius, 0);
-  ctx.closePath();
-  ctx.fill();
-  ctx.restore();
-}
-
-function drawRoundedRect(x, y, width, height, radius) {
-  const r = Math.min(radius, width / 2, height / 2);
-  ctx.beginPath();
-  ctx.moveTo(x + r, y);
-  ctx.lineTo(x + width - r, y);
-  ctx.quadraticCurveTo(x + width, y, x + width, y + r);
-  ctx.lineTo(x + width, y + height - r);
-  ctx.quadraticCurveTo(x + width, y + height, x + width - r, y + height);
-  ctx.lineTo(x + r, y + height);
-  ctx.quadraticCurveTo(x, y + height, x, y + height - r);
-  ctx.lineTo(x, y + r);
-  ctx.quadraticCurveTo(x, y, x + r, y);
-  ctx.closePath();
-  ctx.fill();
-}
-
-function isNear(playerEntity, object, padding) {
-  return (
-    playerEntity.x < object.x + object.width + padding &&
-    playerEntity.x + playerEntity.width > object.x - padding &&
-    playerEntity.y < object.y + object.height + padding &&
-    playerEntity.y + playerEntity.height > object.y - padding
-  );
-}
-
-function gainExperience(amount) {
-  playerStats.exp += amount;
-  let leveledUp = false;
-  let levelsGained = 0;
-  while (playerStats.exp >= playerStats.maxExp) {
-    playerStats.exp -= playerStats.maxExp;
-    playerStats.level += 1;
-    const pointsEarned = getStatPointsForLevel(playerStats.level);
-    playerStats.statPoints = (playerStats.statPoints ?? 0) + pointsEarned;
-    playerStats.maxExp = getExpForNextLevel(playerStats.level);
-    leveledUp = true;
-    levelsGained += 1;
-  }
-  if (levelsGained > 0) {
-    applyAttributeScaling(playerStats);
-  }
-  updateRankFromLevel();
-  syncActiveAccountProgress();
-  if (leveledUp) {
-    audio.playEffect("levelUp");
-  }
-  ui.refresh(playerStats);
-  refreshMissionDisplay();
-  syncMiniGameProfile();
-  return leveledUp;
-}
-
-function clamp(value, min, max) {
-  return Math.max(min, Math.min(max, value));
-}
-
-function showMessage(input, duration, meta = {}) {
-  const payload = typeof input === "string" ? { text: input } : { ...input };
-  const resolvedDuration =
-    typeof payload.duration === "number"
-      ? payload.duration
-      : typeof duration === "number"
-        ? duration
-        : 0;
-  const text = payload.text ?? meta.text ?? "";
-  const channel = payload.channel ?? meta.channel ?? "mission";
-  const author =
-    payload.author ??
-    meta.author ??
-    (channel === "friend" ? "Crewmate" : defaultMissionAuthor);
-  const animate =
-    typeof payload.animate === "boolean"
-      ? payload.animate
-      : typeof meta.animate === "boolean"
-        ? meta.animate
-        : resolvedDuration > 0;
-  const silent =
-    typeof payload.silent === "boolean"
-      ? payload.silent
-      : typeof meta.silent === "boolean"
-        ? meta.silent
-        : !animate;
-  const log =
-    typeof payload.log === "boolean"
-      ? payload.log
-      : typeof meta.log === "boolean"
-        ? meta.log
-        : animate;
-
-  ui.setMessage(text, {
-    ...meta,
-    ...payload,
-    duration: resolvedDuration,
-    channel,
-    author,
-    animate,
-    silent,
-    log
-  });
-
-  if (messageTimerId) {
-    clearTimeout(messageTimerId);
-    messageTimerId = 0;
-  }
-  if (resolvedDuration > 0) {
-    messageTimerId = window.setTimeout(() => {
-      ui.setMessage(defaultMessage, {
-        silent: true,
-        animate: false,
-        log: false
-      });
-      messageTimerId = 0;
-    }, resolvedDuration);
-  }
-}
-
-if (typeof window !== "undefined") {
-  ensureMiniGameEntryPointAvailability();
-  window.astrocatLobby = window.astrocatLobby ?? {};
-  window.astrocatLobby.openMiniGame = openMiniGame;
-  window.astrocatLobby.closeMiniGame = closeMiniGame;
-  window.astrocatLobby.pushTransmission = (payload) => {
-    if (!payload || typeof payload !== "object") {
-      return;
-    }
-    const duration =
-      typeof payload.duration === "number" ? payload.duration : 5200;
-    showMessage(
-      {
-        ...payload,
-        duration,
-        animate: payload.animate ?? true,
-        silent: payload.silent ?? false,
-        log: payload.log ?? true
-      },
-      duration
-    );
-  };
-  window.astrocatLobby.pushFriendPost = (author, text, duration = 5200) => {
-    if (!text) {
-      return;
-    }
-    showMessage(
-      {
-        text,
-        author: author && typeof author === "string" ? author : "Crewmate",
-        channel: "friend",
-        duration,
-        animate: true,
-        silent: false,
-        log: true
-      },
-      duration
-    );
-  };
-
-  miniGameOrigin = computeMiniGameOrigin(miniGameEntryPoint);
-
-  const formatRunDuration = (timeMs) => {
-    const totalSeconds = Math.max(0, Math.round(timeMs / 1000));
-    const minutes = Math.floor(totalSeconds / 60);
-    const seconds = totalSeconds % 60;
-    return `${String(minutes).padStart(2, "0")}:${String(seconds).padStart(2, "0")}`;
-  };
-
-  window.addEventListener("message", (event) => {
-    if (!event || event.origin !== miniGameOrigin) {
-      return;
-    }
-    const data = event.data;
-    if (!data || typeof data !== "object") {
-      return;
-    }
-    const { type, payload } = data;
-    if (type === "astrocat:minigame-transmission") {
-      if (
-        payload &&
-        window.astrocatLobby &&
-        typeof window.astrocatLobby.pushTransmission === "function"
-      ) {
-        window.astrocatLobby.pushTransmission(payload);
-      }
-      return;
-    }
-    if (type === "astrocat:minigame-run") {
-      if (!payload || typeof payload !== "object") {
-        return;
-      }
-      const summary = { ...payload };
-      const normalizedXp = Math.max(
-        0,
-        Math.round(Number.isFinite(summary.xpAward) ? summary.xpAward : Number(summary.xpAward) || 0)
-      );
-      const leveledUp = gainExperience(normalizedXp);
-
-      const playerName =
-        typeof summary.player === "string" && summary.player.trim().length
-          ? summary.player.trim()
-          : "Pilot";
-      const formattedScore = Number.isFinite(summary.score)
-        ? Math.round(summary.score).toLocaleString()
-        : "0";
-      const bestStreak = Number.isFinite(summary.bestStreak)
-        ? Math.max(0, Math.round(summary.bestStreak))
-        : 0;
-      const timeMs = Number.isFinite(summary.timeMs) ? summary.timeMs : 0;
-      const formattedTime =
-        typeof summary.formattedTime === "string" && summary.formattedTime.length
-          ? summary.formattedTime
-          : formatRunDuration(timeMs);
-      const xpLine = normalizedXp > 0 ? ` +${normalizedXp} XP` : "";
-      let messageText = `${playerName} logged ${formattedScore} pts in ${formattedTime} (x${bestStreak} streak).${xpLine}`;
-      const highlightParts = new Set();
-      if (bestStreak >= 12) {
-        const comboMission = completeMission("mission-combo-challenge");
-        if (comboMission.completed) {
-          const xpAward = comboMission.mission?.xp ?? 0;
-          const title = comboMission.mission?.title ?? "Combo Celebration";
-          highlightParts.add(`Mission Complete: ${title}! +${xpAward} EXP.`);
-          if (comboMission.leveledUp) {
-            highlightParts.add(`Level up! You reached level ${playerStats.level}.`);
-          }
-        } else if (comboMission.locked) {
-          const required = Math.max(1, comboMission.mission?.requiredLevel ?? 1);
-          highlightParts.add(`Reach Level ${required} to log combo triumphs.`);
-        }
-      }
-      if (leveledUp) {
-        highlightParts.add(`Level up! You reached level ${playerStats.level}.`);
-      }
-      if (highlightParts.size > 0) {
-        messageText += ` ${Array.from(highlightParts).join(" ")}`;
-      }
-      showMessage(
-        {
-          text: messageText,
-          channel: "mission",
-          duration: 6400,
-          animate: true,
-          silent: false,
-          log: true
-        },
-        6400
-      );
-    }
-  });
 }
 
 function createOnboardingExperience(config = {}) {
